API Documentation

Parsing

A parser is created from a Lark grammar by calling the Lark constructor. Parsing is initiated by calling parse.

Missing docstring.

Missing docstring for Lark(grammar::String;options...). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Lark(grammar::IOStream,source;options...). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Lerche.open(grammar_filename;rel_to=nothing,options...). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Lerche.parse(l::Lark,text;start=nothing,on_error=nothing). Check Documenter's build log for details.

Working with the parse tree

transform transforms the parse tree according to rules defined by the user using @rule and @inline_rule macros. Tokens will also be processed if visit_tokens returns true for that transformer type. This will significantly slow down parse tree processing.

Missing docstring.

Missing docstring for @rule s. Check Documenter's build log for details.

Missing docstring.

Missing docstring for @inline_rule s. Check Documenter's build log for details.

Lerche.Transformer — Type

Transformers visit each node of the tree, and run the appropriate method on it according to the node's data.

The methods (provided by the user) are dispatched according to the type and ``tree.data``.
The returned value replaces the old one in the structure.

They work bottom-up (or depth-first), starting with the leaves and ending at the root of the tree.
Transformers can be used to implement map & reduce patterns. Because nodes are reduced from leaf to root,
at any point the callbacks may assume the children have already been transformed (if applicable).

``Transformer`` can do anything ``Visitor`` can do, but because it reconstructs the tree,
it is slightly less efficient. It can be used to implement map or reduce patterns.

All these classes implement the transformer interface:

- ``Transformer`` - Recursively transforms the tree. This is the one you probably want.
- ``Transformer_InPlace`` - Non-recursive. Changes the tree in-place instead of returning new instances
- ``Transformer_InPlaceRecursive`` - Recursive. Changes the tree in-place instead of returning new instances

Traits: ``.

NOTE: A transformer without methods essentially performs a non-memoized deepcopy.

Missing docstring.

Missing docstring for Transformer_InPlace. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Transformer_InPlaceRecursive. Check Documenter's build log for details.

Lerche.visit_tokens — Method

visit_tokens (::Transformer) = true

Should the transformer visit tokens in addition to rules. Setting this to false is an order of magnitude faster. Defaults to true. (For processing ignored tokens, use the $lexer_callbacks$ options)

Missing docstring.

Missing docstring for transform(tr::Transformer,tree). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Visitor. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Visitor_Recursive. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Interpreter. Check Documenter's build log for details.

Missing docstring.

Missing docstring for visit(v::Visitor,tree). Check Documenter's build log for details.